tests: mark old bsd fdisk test as deprecated
It seems better to use hexdump rather than md5sum, but it means that
we have to gather hexdumps of the all possible BSD variants. For this
purpose will be introduced a new bsd fdisk test and to verify the
new hexdumps we can use this old test as both tests are exactly the
same.

Signed-off-by: Karel Zak <kzak@redhat.com>