Fail like the shell when execvp fails
Commands executing another command (e.g. taskset) should fail in a
consistent and detectable way when the other command fail to execute.
Change most of the commands that used to fail with EXIT_FAILURE to fail
with the special exit code 127, like the shell.

For example, running the fuser command via taskset when fuser command is
not installed is failing now like this:

    $ taskset -c 1 fuser /path/to/file; echo $?
    taskset: failed to execute fuser: No such file or directory
    1

This failure looks like a real failure of the fuser command:

    $ taskset -c 1 fuser /path/to/file; echo $?
    Specified filename /path/to/file does not exist.
    1

Based on the exit code, a management program may do the wrong thing,
leading to catastrophic results. Detecting such failure requires fragile
parsing of standard error.

With this patch this failure is easy to detect:

    $ taskset -c 1 fuser /path/to/file; echo $?
    taskset: failed to execute fuser: No such file or directory
    127

The return code seems to be undocumented for most commands, except the
su command, promising to exit with exit code 126 if a command cannot be
executed, and 127 if a command was not found. However, the actual
command always exit with exit code 1. This patch implement the
documented behavior.

Signed-off-by: Nir Soffer <nsoffer@redhat.com>