libmount: ignore double slashes in path start
With NFSv4, the path appearing in /proc/mounts may further differ from
the one for which it was asked (say in /etc/fstab) and /run/mount/utab.
That causes problems at least for applying userspace-stored flags (like
"user").

server exports | client mounts | /proc/mounts | utab        | ->
               |               |              |             |
/tmp           | euler://tmp   | euler:/tmp   | euler://tmp | ERR
               | euler:/tmp    | euler:/tmp   | euler:/tmp  | OK
               |               |              |             |
/              | euler://tmp   | euler://tmp  | euler://tmp | OK
               | euler:/tmp    | euler://tmp  | euler:/tmp  | ERR

It surely would be better if kernel was behaving differently, yet after
discussing that with NFS people it sounds unlikely to be fixed in the
near future.

This patch modifies path comparison to treat paths as identical, if they
contain ':' and directly after it one has / and the other // - and the
rest is the same, except for the trailing slashes.

Signed-off-by: Ales Novak <alnovak@suse.cz>