lscpu: update s390-lpar-drawer testcase
I noticed that the "cpu" line within /proc/cpuinfo has a different
meaning on powerpc and sparc than on s390.

On s390 it represents the cpu number while on powerpc and sparc it
represent the model name. Given that the s390 kernel patch which adds
the "cpu" line to /proc/cpuinfo hasn't been merged into the kernel
yet, I changed that line to "cpu number" to avoid the unnecessary
conflict.

Therefore this patch updates the testcase, which mainly changes the
contents of /proc/cpuinfo.

Signed-off-by: Heiko Carstens <heiko.carstens@de.ibm.com>